generate PfsConfig and use that to generate the spectra
We need to produce a PfsConfig along with the observation (for the
pipeline). Moving towards generating the spectra from the PfsConfig
(or a PfiDesign), which is kinda how the instrument will operate.
The PfsConfig specifies catId,objId for each fiber, which is used
with a SpectrumLibrary to generate the spectra which are fed into
the simulator. This now includes a science spectrum, which is
currently hardwired to an example galaxy spectrum from the 1D simulator
(but this can easily be extended to read a variety of spectra based on
the objId).
Intertwined with this, changed the output to use the raw file format
in the datamodel (PFFAxxxxxxyz.fits).
Also some cleanup and added a "--pdb" command-line argument for debugging.

Includes