redesign Spectrum class for proper flux scaling
We've been simulating a spectrum that's simply interpolated at the
wavelength of interest, which is fine if the lines are well-sampled;
however, narrow lines might disapppear. In order to account for this,
we need to integrate the spectrum over the wavelength range of each
spot. As part of this, being more careful about units and distinguishing
between flux and flux density. Continuum sources specify flux density
in nJy, and integrate those to provide a flux in W/m^2 (could have been
cgs, but I had to choose something, and SI units seemed like a reasonable
choice). Line sources provide a flux in W/m^2 (effectively, they have
already integrated over the line profile).

Reimplemented all the Spectrum subclasses with the new design.

Added functions that calculate the appropriate flux or flux density given
a desired photon flux, to help with setting reasonable levels on the
"artificial" spectra.