implement the sky swindle
We only add the noise from the sky, not the sky itself.
This is controlled by a command-line option.

The sky swindle is implemented in Exposure.readout.
Currently, the simulator has to put the spots down twice; we may or may
not be able to avoid this by being a bit more clever, but this is left
to later.