add dirty solution for shifting the slit
Shifting the slit (through --xoffset and --yoffset) didn't work; looks
like it worked only with the previous spot data, but not with the modern
JEG spots.

I don't see how to shift the slit properly (it should be nonlinear due
to the optics), and so I'm implementing a dirty fix: simply perform a
linear shift on the detector. This is only needed for building flats, so
it's not a big deal, and will let me proceed; the exact fix can be
pursued later.