remove the sky swindle
As currently implemented, the sky swindle requires the pipeline to read
a variance plane from the simulator, because the noised flux in the
spectrum (flux+sky fluctuates low, then subtract the sky) can be negative.
The alternative is to have the pipeline subtract the known sky, which
is easier, and doesn't require any special handling in the simulator
(except for adding --allOutputs).