SkyModel: refactor and add extinction
Using Spectrum rather than splines (a TableSpectrum is a kind of a spline,
it's just a different representation). I want to be careful with the
resolution: can't just multiply tables/splines at different resolution.

This includes a new Arm enum class for standardising how the arm
is represented (band vs arm).

Extinction data come from the data recently added to drp_instdata,
provided by Nicholas Mondrik. Note that we now use the second-to-last
column of the instrument throughput data, which contains the instrument
only, without the atmosphere.