allow construction of multiple identical images at once
We often construct multiple images that are identical (e.g., biases,
darks, flats) apart from the readout characteristics (read noise, bias,
CRs). There's no need to do the calculations to put the fibers down
multiple times because they're the same for each. Instead, we clone the
image with the fibers down, and perform the readout on that, allowing us
to repeat this multiple times to write out multiple images.