makeSimRun: allow general parallelism
We have previously parallelised across detectors by iterating
over detectors, launching jobs in the background and then
waiting for those to complete. This limits the parallelism to
the number of detectors only.

This commit adds a general parallelism, with the number of
cores set by a command-line parameter. We first construct a
list of the commands to run, and then use xargs to
run them in parallel.