makeSimRun: reduce number of exposures
to mimimise the size of the dataset.

Removed half the biases, half the darks, a third of the flats and
half the arcs. In an ideal world, we'd keep them, but we need to
cut back on the size, and this dataset isn't for getting an idea
of quality, but to make sure nothing's broken.