makeDetectorMap: workaround non-physical features in wavelength solution
The detectorMaps coming from the simulator have non-physical features at
either end, possibly due to spline artifacts. Jim Gunn says, "There is
nothing physical which could cause this." We suspect the problem is in the
splining of the optical model. This change implements a workaround, which
is to drop the second, third, third-to-last and second-to-last knots for
the wavelength solution from the detectorMap.