write DATE-OBS as some base time + visit*minutes
This ensures that a data set's DATE-OBS will always sort in the same
order as the visit number (which is important for calibs). Previously,
the DATE-OBS was set to the time at which the image is generated, which
can be out of visit order when we're generating images in parallel.

This steals the 'date' argument to makeSim, which wasn't doing anything,
and makes it the base time.