Moved hard-coded mapping of catId to catalog details (name, location) to a separate YAML file.
This enables the mapping of the catId and
relative location of the actual catalog object spectra
to be configured separately to the code.

This change also means the introduction
of an additional command-line argument that provides the root location
of the catalog data. This can be different to the location of instrument
data (PSFs etc).