makeDetectorMap: include engineering fibers
We have spots for the science fibers only, but we want to also include
the engineering fibers in the detectorMap (they'll see light, which we
will need to deal with in order to avoid contaminating the science
spectra). In order to do so, we interpolate the positions of neighbouring
fibers.