strip out support for reading AMBRE spectra
We have been using AMBRE spectra to do flux calibration, but now we have
our own spectra (derived from AMBRE, but reformatted) for that. Ensure that
those are used, and remove everything associated with the old way of doing
fluxCal, so we can be assured we're using the new way now.