write NIR data as PFSB files
This moves the capability of the convertPFSAtoPFSB.py script from
PIPE2D-1088 into the simulator, so there's no need to post-process
the simulator output. We write the NIR image as an up-the-ramp
sequence with two reads: the first zero, and the second the simulated
image. The reference for the second is the bias level (no noise, or
it would subtract out the noise in the image). The HDUs are
compressed, so it shouldn't add a large disk overhead.